+++ BodyPositiveZumba/Home/Home/HomeDomain/HomeCoordinator.swift
@@ -11,7 +11,6 @@ import SwiftUI
 protocol HomeCoordinator: ScreenCoordinator {
     
     func home_viewAbout()
-    func home_viewSideDrawer()
     func home_viewClasses()
     func home_viewJoinNow()
