+++ BodyPositiveZumba/Home/Home/HomeUI/HomeScreen.swift
@@ -10,17 +10,21 @@ import SwiftUI
 struct HomeScreen: View {
 
     @ObservedObject private var viewModel: HomeViewModel
+    private var sideDrawerCoordinator: SideDrawerCoordinator
 
     public init(
+        sideDrawerCoordinator: SideDrawerCoordinator,
         viewModel: HomeViewModel
     ) {
+        self.sideDrawerCoordinator = sideDrawerCoordinator
         self._viewModel = ObservedObject(wrappedValue: viewModel)
     }
 
     var body: some View {
         HomeView(
-            coordinator: viewModel.sideDrawerCoordinator
-        ) { action in
+            coordinator: sideDrawerCoordinator
+        ) {
+            action in
             switch action {
             }
         }
